requirements.txt
pybit
pycryptodome
python-dotenv